Move unit tests into `tests.rs` files
To not reduce the code coverage involuntarily
