+++ codecov.yml
@@ -1,2 +1,3 @@
 ignore:
+  - "**/tests.rs"
   - "**/tests_*.rs"
